# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"CurrentStatus": {
				"enum": [
					"passive",
					"rejected",
					"pending"
				],
				"type": "string"
			},
			"AppliedForRole": {
				"enum": [
					"default",
					"artist",
					"curator",
					"gallery"
				],
				"type": "string"
			},
			"Status": {
				"properties": {
					"id": {
						"type": "string"
					},
					"currentStatus": {
						"$ref": "#/components/schemas/CurrentStatus"
					},
					"appliedForRole": {
						"$ref": "#/components/schemas/AppliedForRole"
					},
					"appliedForCategory": {
						"type": "string"
					}
				},
				"required": [
					"appliedForCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterResponseModel": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"approvedBy": {
						"type": "string"
					},
					"instagram": {
						"type": "string"
					},
					"mirrorxyz": {
						"type": "string"
					},
					"linkedIn": {
						"type": "string"
					},
					"facebook": {
						"type": "string"
					},
					"twitter": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"nonce": {
						"type": "string"
					},
					"bannerHash": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"totalSales": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string"
					},
					"isVerified": {
						"type": "boolean"
					},
					"biography": {
						"type": "string"
					},
					"publicAddress": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"totalSales",
					"role",
					"isVerified",
					"publicAddress",
					"lastname",
					"name",
					"email"
				],
				"type": "object"
			},
			"RegisterRequestModel": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/Status"
					},
					"approvedBy": {
						"type": "string"
					},
					"instagram": {
						"type": "string"
					},
					"mirrorxyz": {
						"type": "string"
					},
					"linkedIn": {
						"type": "string"
					},
					"facebook": {
						"type": "string"
					},
					"twitter": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"nonce": {
						"type": "string"
					},
					"bannerHash": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"totalSales": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string"
					},
					"isVerified": {
						"type": "boolean"
					},
					"biography": {
						"type": "string"
					},
					"publicAddress": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"required": [
					"totalSales",
					"role",
					"isVerified",
					"publicAddress",
					"lastname",
					"name",
					"email"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "clean-architecture-demo",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/authentication/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/RegisterResponseModel"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequestModel"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}